requirements.txt
requests>=2.31.0,<2.32.0
pillow>=10.2.0,<11.0.0
